code factorization between vtkvolume and vtkslicer
